feat: Enhance OPFS client and region file management
- Implemented a new stats retrieval feature in OpfsClient and opfs.worker.ts to monitor slot usage and memory statistics.
- Refactored the OpfsClient to use a fixed-size array for pending operations, improving performance and memory management.
- Added LRU caching mechanism for region files in opfs.worker.ts to limit the number of open regions and optimize memory usage.
- Introduced new block types (CoralBlock, PalmTrunk, PalmLeaves) in BlockType.ts for expanded world features.
- Updated RegionFile to utilize pre-allocated typed arrays for compaction, reducing heap allocations and improving performance.
- Modified AtlasBuilder.html to enable image smoothing for better texture quality.
- Cleaned up WorldStorage to ensure proper removal of old data files and directories.

+++ src/code/World/MeshPipeline/core/WaterPipeline.ts
@@ -45,7 +45,9 @@ export interface WaterSampleGrid {
 
 // PERF: Version-stamp tracking array — reused across builds to avoid
 // allocating + zeroing a new Uint8Array per water mesh build.
-let _waterUsedStamp: Uint8Array | null = null;
+// Uses Uint16Array to avoid overflow (Uint8Array wraps at 255, causing
+// version collision with the "unused" sentinel 0 and duplicate quads).
+let _waterUsedStamp: Uint16Array | null = null;
 let _waterVersion = 0;
 
 export function buildWaterMesh(
@@ -60,7 +62,7 @@ export function buildWaterMesh(
 	// PERF: Use version stamp instead of allocating + zeroing a Uint8Array per build.
 	// Each build increments the stamp; cells are "used" if their stamp matches.
 	if (!_waterUsedStamp || _waterUsedStamp.length < samples.length) {
-		_waterUsedStamp = new Uint8Array(samples.length);
+		_waterUsedStamp = new Uint16Array(samples.length);
 	}
 	_waterVersion++;
 	const version = _waterVersion;
